enable Ads in the tvOS sample
* extract common parts into a shared folder
