+++ Example/BitmovinConvivaAnalytics/ViewController.swift
@@ -26,6 +26,18 @@ class ViewController: UIViewController {
     let convivaCustomerKey: String = "YOUR-CONVIVA-CUSTOMER-KEY"
     var convivaGatewayString: String?
 
+    var vodSourceConfig: SourceConfig {
+        var sourceString = "https://bitmovin-a.akamaihd.net/content/art-of-motion_drm/m3u8s/11331.m3u8"
+        if let streamString = streamUrlTextField.text,
+           URL(string: streamString) != nil {
+            sourceString = streamString
+        }
+
+        let sourceConfig = SourceConfig(url: URL(string: sourceString)!, type: .hls)
+        sourceConfig.title = "Art of Motion"
+        return sourceConfig
+    }
+
     override func viewDidLoad() {
         super.viewDidLoad()
 
@@ -46,7 +58,8 @@ class ViewController: UIViewController {
 
     func setupBitmovinPlayer() {
         // Setup Player
-        player = PlayerFactory.create(playerConfig: playerConfig)
+        let playerConfig = buildDefaultPlayerConfig(enableAds: adsSwitch.isOn)
+        player = PlayerFactory.createPlayer(playerConfig: playerConfig)
 
         let convivaConfig = ConvivaConfiguration()
 
@@ -82,53 +95,7 @@ class ViewController: UIViewController {
         playerView?.autoresizingMask = [.flexibleHeight, .flexibleWidth]
         playerUIView.addSubview(playerView!)
 
-        player?.load(source: SourceFactory.create(from: vodSourceConfig))
-    }
-
-    var playerConfig: PlayerConfig {
-        let playerConfig = PlayerConfig()
-        if adsSwitch.isOn {
-            playerConfig.advertisingConfig = adConfig
-        }
-
-        let playbackConfig = PlaybackConfig()
-        playbackConfig.isAutoplayEnabled = true
-        playbackConfig.isMuted = true
-        playerConfig.playbackConfig = playbackConfig
-        return playerConfig
-    }
-
-    var vodSourceConfig: SourceConfig {
-        var sourceString = "https://bitmovin-a.akamaihd.net/content/art-of-motion_drm/m3u8s/11331.m3u8"
-        if let streamString = streamUrlTextField.text,
-           URL(string: streamString) != nil {
-            sourceString = streamString
-        }
-
-        let sourceConfig = SourceConfig(url: URL(string: sourceString)!, type: .hls)
-        sourceConfig.title = "Art of Motion"
-        return sourceConfig
-    }
-
-    var adConfig: AdvertisingConfig {
-        // swiftlint:disable:next line_length
-        let adTagVastSkippable = "https://pubads.g.doubleclick.net/gampad/ads?sz=640x480&iu=/124319096/external/single_ad_samples&ciu_szs=300x250&impl=s&gdfp_req=1&env=vp&output=vast&unviewed_position_start=1&cust_params=deployment%3Ddevsite%26sample_ct%3Dskippablelinear&correlator=1"
-        // swiftlint:disable:next line_length
-        let adTagVast1 = "https://pubads.g.doubleclick.net/gampad/ads?sz=640x480&iu=/124319096/external/single_ad_samples&ciu_szs=300x250&impl=s&gdfp_req=1&env=vp&output=vast&unviewed_position_start=1&cust_params=deployment%3Ddevsite%26sample_ct%3Dlinear&correlator=2"
-        // swiftlint:disable:next line_length
-        let adTagVast2 = "https://pubads.g.doubleclick.net/gampad/ads?sz=640x480&iu=/32573358/2nd_test_ad_unit&ciu_szs=300x100&impl=s&gdfp_req=1&env=vp&output=vast&unviewed_position_start=1&url=[referrer_url]&description_url=[description_url]&correlator=3"
-
-        let adSource1 = AdSource(tag: URL(string: adTagVastSkippable)!, ofType: .ima)
-        let adSource2 = AdSource(tag: URL(string: adTagVast1)!, ofType: .ima)
-        let adSource3 = AdSource(tag: URL(string: adTagVast2)!, ofType: .ima)
-
-        let preRoll = AdItem(adSources: [adSource1], atPosition: "pre")
-        let midRoll = AdItem(adSources: [adSource2], atPosition: "20%")
-        let postRoll = AdItem(adSources: [adSource3], atPosition: "post")
-
-        let adConfig = AdvertisingConfig(schedule: [preRoll, midRoll, postRoll])
-
-        return adConfig
+        player?.load(source: SourceFactory.createSource(from: vodSourceConfig))
     }
 
     // MARK: - actions
